--- examples/model/README.md
@@ -1,3 +1,0 @@
-## Model Binaries
-
-The binaries located in this directory are the components of a trained rllib model. These are related to the `examples/rllib.py` example script. Results from `examples/rllib.py` are loaded and written to this directory.
