+++ smarts/core/sumo_traffic_simulation.py
@@ -30,10 +30,6 @@ from shapely.affinity import rotate as shapely_rotate
 from shapely.geometry import Polygon
 from shapely.geometry import box as shapely_box
 
-from smarts.core.utils.sumo import SUMO_PATH, traci
-import traci.constants as tc
-from traci.exceptions import FatalTraCIError, TraCIException
-
 from smarts.core import gen_id
 from smarts.core.colors import SceneColors
 from smarts.core.coordinates import Dimensions, Heading, Pose
@@ -43,6 +39,9 @@ from smarts.core.utils import networking
 from smarts.core.utils.logging import suppress_output
 from smarts.core.vehicle import VEHICLE_CONFIGS, VehicleState
 
+from smarts.core.utils.sumo import SUMO_PATH, traci # isort:skip
+from traci.exceptions import FatalTraCIError, TraCIException # isort:skip
+import traci.constants as tc # isort:skip
 
 class SumoTrafficSimulation(Provider):
     """
