+++ smarts/examples/__init__.py
