+++ smarts/examples/egoless.py
@@ -2,8 +2,8 @@ import logging
 
 import gym
 
-from examples.argument_parser import default_argument_parser
 from smarts.core.utils.episodes import episodes
+from smarts.examples.argument_parser import default_argument_parser
 
 logging.basicConfig(level=logging.INFO)
