+++ smarts/examples/model/README.md
@@ -1,0 +1,3 @@
+## Model Binaries
+
+The binaries located in this directory are the components of a trained rllib model. These are related to the `smarts/examples/rllib.py` example script. Results from `smarts/examples/rllib.py` are loaded and written to this directory.
