+++ smarts/examples/model/variables/variables.data-00000-of-00001
