+++ smarts/examples/model/variables/variables.index
