+++ smarts/examples/multi_instance.py
@@ -11,15 +11,15 @@ import torch
 try:
     import ray
 except Exception as e:
-    from examples import RayException
+    from smarts.examples import RayException
 
     raise RayException.required_to("multi_instance.py")
 
 
-from examples.argument_parser import default_argument_parser
 from smarts.core.agent import Agent, AgentSpec
 from smarts.core.agent_interface import AgentInterface, AgentType
 from smarts.core.utils.episodes import episodes
+from smarts.examples.argument_parser import default_argument_parser
 
 logging.basicConfig(level=logging.INFO)
