+++ smarts/examples/observation_collection_for_imitation_learning.py
@@ -1,18 +1,18 @@
 import logging
-import pickle
 import os
+import pickle
 from typing import Any, Dict, Sequence
 
 from envision.client import Client as Envision
-from examples.argument_parser import default_argument_parser
 from smarts.core.agent import AgentSpec
 from smarts.core.agent_interface import AgentInterface, AgentType
+from smarts.core.controllers import ControllerOutOfLaneException
 from smarts.core.scenario import Scenario
 from smarts.core.sensors import Observation
 from smarts.core.smarts import SMARTS
 from smarts.core.sumo_traffic_simulation import SumoTrafficSimulation
 from smarts.core.utils.math import radians_to_vec
-from smarts.core.controllers import ControllerOutOfLaneException
+from smarts.examples.argument_parser import default_argument_parser
 
 logging.basicConfig(level=logging.INFO)
