+++ smarts/examples/ros/.catkin_workspace
