+++ smarts/examples/ros/.gitignore
