+++ smarts/examples/ros/src/smarts_ros/msg/EntitiesStamped.msg
