+++ smarts/examples/ros/src/smarts_ros/msg/SmartsReset.msg
