+++ smarts/examples/ros/src/smarts_ros/package.xml
