+++ smarts/examples/tools/pybullet_linear_tire_example.py
@@ -1,19 +1,19 @@
 import math
 from pathlib import Path
 from unittest import mock
-import numpy as np
-import matplotlib.pyplot as plt
 
+import matplotlib.pyplot as plt
+import numpy as np
+
+from smarts.core.chassis import AckermannChassis
 from smarts.core.controllers.actuator_dynamic_controller import (
     ActuatorDynamicController,
     ActuatorDynamicControllerState,
 )
-from smarts.core.coordinates import Pose, Heading
-from smarts.core.vehicle import Vehicle
-from smarts.core.chassis import AckermannChassis
+from smarts.core.coordinates import Heading, Pose
 from smarts.core.utils import pybullet
 from smarts.core.utils.pybullet import bullet_client as bc
-
+from smarts.core.vehicle import Vehicle
 
 TIMESTEP_SEC = 1 / 240
 # TIMESTEP_SEC=0.0005
