+++ smarts/examples/tools/pybullet_sumo_orientation_example.py
@@ -1,19 +1,18 @@
 import math
+import multiprocessing as mp
 import random
 from pathlib import Path
 from unittest import mock
-import multiprocessing as mp
 
 import numpy as np
 
+from smarts.core.chassis import BoxChassis
 from smarts.core.coordinates import Heading, Pose
 from smarts.core.scenario import Scenario
-from smarts.core.vehicle import VEHICLE_CONFIGS, Vehicle, VehicleState
-from smarts.core.chassis import BoxChassis
+from smarts.core.sumo_traffic_simulation import SumoTrafficSimulation
 from smarts.core.utils import pybullet
 from smarts.core.utils.pybullet import bullet_client as bc
-from smarts.core.sumo_traffic_simulation import SumoTrafficSimulation
-
+from smarts.core.vehicle import VEHICLE_CONFIGS, Vehicle, VehicleState
 
 TIMESTEP_SEC = 1 / 240
 INITIAL_KINEMATICS_VEHICLES = 50
