+++ smarts/examples/tools/regression_rllib.py
@@ -14,15 +14,15 @@ try:
     from ray.rllib.models import ModelCatalog
     from ray.rllib.utils import try_import_tf
 except Exception as e:
-    from examples import RayException
+    from smarts.examples import RayException
 
     raise RayException.required_to("regression_rllib.py")
 
-from examples.rllib_agent import TrainingModel
 from smarts.core.agent import Agent, AgentSpec
+from smarts.core.agent_interface import AgentInterface, AgentType
 from smarts.env.custom_observations import lane_ttc_observation_adapter
 from smarts.env.rllib_hiway_env import RLlibHiWayEnv
-from smarts.core.agent_interface import AgentInterface, AgentType
+from smarts.examples.rllib_agent import TrainingModel
 
 HORIZON = 5000
