+++ smarts/examples/tools/sumo_multi_clients.py
@@ -4,7 +4,7 @@ import subprocess
 import threading
 import time
 
-from smarts.core.utils.sumo import sumolib, traci, SUMO_PATH
+from smarts.core.utils.sumo import SUMO_PATH, sumolib, traci
 
 PORT = 8001
