+++ smarts/examples/tools/tire_parameters.yaml
