+++ smarts/examples/trajectory_tracking_agent.py
@@ -2,10 +2,10 @@ import logging
 
 import gym
 
-from examples.argument_parser import default_argument_parser
 from smarts.core.agent import Agent, AgentSpec
 from smarts.core.agent_interface import AgentInterface, AgentType
 from smarts.core.utils.episodes import episodes
+from smarts.examples.argument_parser import default_argument_parser
 
 logging.basicConfig(level=logging.INFO)
