+++ zoo/policies/cross-rl-agent/cross_rl_agent/train/run_test.py
@@ -19,7 +19,7 @@
 # THE SOFTWARE.
 # The author of this file is: https://github.com/mg2015started
 
-# The following test was modified from examples/multi_instance.py
+# The following test was modified from smarts/examples/multi_instance.py
 
 import argparse
 import logging
